Move shared patch validation operations into module
