+++ packages/two_percent/lib/two_percent.rb
@@ -5,10 +5,10 @@ require "aether_observatory"
 require "two_percent/version"
 require "two_percent/configuration"
 require "two_percent/domain"
+require "two_percent/validates_user_groups_patch"
 require "two_percent/bulk_processor"
 require "two_percent/scim"
 require "two_percent/syncable"
-
 module TwoPercent
   # Custom exception for SCIM RFC 7643 read-only attribute violations
   class ReadOnlyAttributeError < StandardError; end
